Amended creating cursor above/below to skip lines, added create_cursors_around
If the current cursor is not at the start of the line, when the user creates a cursor above or below it, empty lines will be skipped.  This is a lot more ergonomic; almost always this is what the user wants to do.
The old behaviour is preserved when the cursor is at the start of the line (for instance, when on an empty line).

create_cursors_around:
Cursors will be added to every contiguous line above and below the current cursor as long as the line in question reaches to the cursor column.