Add patch/diff file highlighting
Added patch/diff file highlighting for easy at a glance viewing.

Does not try to highlight based on language, just a simple green for additions
and red for deletions, plus some extra colours for metadata.

(This could be changed in the future if we did want to highlight per language
since the regions I added could stay the same and just change the token colours...
That would involve needing to detect language based on the file extensions of
the diff'd files though which might be kind of slow, especially for large
patches that involve many files.)