Bugfix in change_case_of_selected_text
change_case_of_selected_text was modifying buffer directly in place.
It wasn't marking file as dirty and it wasn't working with undo/redo.
replace_range called at the end of this function wasn't really doing
anything.
This is fixed by introducing temporary string copy that is modified
and passed to replace_range.